Make Cython testing actually work
Previously, the cython stage in Travis would build (I think) but not use the .so files during
testing, because of how py.test adds the current working directory to sys.path - the
.so files would be installed in the tox virtualenv, but python would import faust
from the faust/ directory under the git clone.

This made tests that would otherwise fail with the Cython implementation pass, as they
were testing the python implementation of modules.

Also adds one test under t/cython that will fail if we can't import a cythonized module,
ensuring we get a proper warning if we have an incorrect test environment